Refactor IPv6 Extension Header Handling
Use an (ordered!) list of the types of IPv6 extension headers that can
be handled to drive their processing. The goal: Make the order in which
extension headers are in the packet on the wire deterministic.

This patch removes special-case code that forces hop-by-hop extension
headers to be handled before all others (RFC8200: "The Hop-by-Hop
Options header, when present, must immediately follow the IPv6 header.")

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>