Move Initial Setup Of IP Packet Length
The calculation of the baseline total length of an IP packet should be
calculated independently of whether there is an application body or not.
In earlier versions of the code, the packet's total length was only
calculated when the body field of the packet was set (if at all). The
problem was masked because the test cases only included code where there
was a body set (and Linux will overwrite that field on a socket where
the IP header is included).

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>